The comment for push
The are the python code files and the created model. The data set is not taken from any source but I created a data set using python code which include:
time, nasal_airflow,apnea_event,hypopnea_event,event,recording as the features to build the model. while creating the data set, to make it realistic I added some noise to the data using the sin function.